Finished reroute of koroks and for Climber's Set
Rerouted some koroks in the bay area north of Mt. Lanayru and south of Zora's Domain. Change the route to get the Climber's Set upgraded to level 2 as quickly as possible. Moved more shrines to the beginning of the route so that the loss of stamina when getting the Master Sword is not a bad. Routed in more fleet-lotus seeds.